Basic UML Diagram added
This diagram displays the basic design of the simulation. The classes: Devices, Link and Packet are all part of the class Network Description. The relation between Devices and Network Cards is that a device can have one or many network cards. Devices may as well have several links, or send/receive several packets. The two protocols, Network Layer and Link Layer, represent respectively IPv4/IPv6 and Ethernet/Wi-Fi. This simulator as mentioned in the brief, should be able to provide a platform for simulating both of them.